+++ crates/nix_rs/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## Unreleased
-
+- **`flake::schema`**
+  - Don't hardcode flake schema types
 - **`env`**:
   - use `whoami` crate to find the current user instead of depending on environment variable `USER`
 - Support Nix 2.20